+++ src/main/resources/_config.properties
@@ -55,6 +55,8 @@ capabilities.appPackage=com.mercadolibre
 capabilities.appActivity=activities.SplashActivity
 capabilities.noReset=true
 capabilities.autoGrantPermissions=true
+capabilities.androidInstallTimeout=5000
+
 
 #uncomment and change udid and deviceName for Android test
 #===================Android -=============#
